+++ sun/sun-api/models/session.go
@@ -31,8 +31,8 @@ type Session struct {
 	DeviceMAC   string    `db:"device_mac" json:"device_mac"`
 	IpAddress   string    `db:"ip_address" json:"ip_address"`
 	Username    string    `db:"username" json:"username"`
-	BytesUp     int       `db:"bytes_up" json:"bytes_up"`
-	BytesDown   int       `db:"bytes_down" json:"bytes_down"`
+	BytesUp     int64     `db:"bytes_up" json:"bytes_up"`
+	BytesDown   int64     `db:"bytes_down" json:"bytes_down"`
 	Duration    int       `db:"duration" json:"duration"`
 	AuthTime    time.Time `db:"auth_time" json:"auth_time"`
 	StartTime   time.Time `db:"start_time" json:"start_time"`
@@ -54,12 +54,12 @@ type Session struct {
 }
 
 type SessionLimit struct {
-	TrafficBytes    int `db:"traffic_bytes" json:"traffic_bytes"`
-	DurationSeconds int `db:"duration_seconds" json:"duration_seconds"`
+	TrafficBytes    int64 `db:"traffic_bytes" json:"traffic_bytes"`
+	DurationSeconds int   `db:"duration_seconds" json:"duration_seconds"`
 }
 type SessionHistoryLimit struct {
-	TrafficBytes    int `db:"traffic_bytes" json:"traffic_bytes"`
-	DurationSeconds int `db:"duration_seconds" json:"duration_seconds"`
+	TrafficBytes    int64 `db:"traffic_bytes" json:"traffic_bytes"`
+	DurationSeconds int   `db:"duration_seconds" json:"duration_seconds"`
 }
 
 func (SessionLimit) TableName() string {
