+++ sun/sun-api/models/user.go
@@ -39,7 +39,7 @@ type User struct {
 	SurveyAuth           bool      `db:"survey_auth" json:"survey_auth"`
 	KbpsDown             int       `db:"kbps_down" json:"kbps_down"`
 	KbpsUp               int       `db:"kbps_up" json:"kbps_up"`
-	MaxNavigationTraffic int       `db:"max_navigation_traffic" json:"max_navigation_traffic"`
+	MaxNavigationTraffic int64     `db:"max_navigation_traffic" json:"max_navigation_traffic"`
 	MaxNavigationTime    int       `db:"max_navigation_time" json:"max_navigation_time"`
 	AutoLogin            bool      `db:"auto_login" json:"auto_login"`
 	ValidFrom            time.Time `db:"valid_from" json:"valid_from"`
@@ -62,9 +62,9 @@ type UserJSON struct {
 	SurveyAuth                bool      `db:"survey_auth" json:"survey_auth"`
 	KbpsDown                  int       `db:"kbps_down" json:"kbps_down"`
 	KbpsUp                    int       `db:"kbps_up" json:"kbps_up"`
-	MaxNavigationTraffic      int       `db:"max_navigation_traffic" json:"max_navigation_traffic"`
+	MaxNavigationTraffic      int64     `db:"max_navigation_traffic" json:"max_navigation_traffic"`
 	MaxNavigationTime         int       `db:"max_navigation_time" json:"max_navigation_time"`
-	MaxNavigationTrafficLimit int       `db:"max_navigation_traffic_limit" json:"max_navigation_traffic_limit"`
+	MaxNavigationTrafficLimit int64     `db:"max_navigation_traffic_limit" json:"max_navigation_traffic_limit"`
 	MaxNavigationTimeLimit    int       `db:"max_navigation_time_limit" json:"max_navigation_time_limit"`
 	AutoLogin                 bool      `db:"auto_login" json:"auto_login"`
 	ValidFrom                 time.Time `db:"valid_from" json:"valid_from"`
