+++ sun/sun-api/models/user_history.go
@@ -40,7 +40,7 @@ type UserHistory struct {
 	SurveyAuth           bool      `db:"survey_auth" json:"survey_auth"`
 	KbpsDown             int       `db:"kbps_down" json:"kbps_down"`
 	KbpsUp               int       `db:"kbps_up" json:"kbps_up"`
-	MaxNavigationTraffic int       `db:"max_navigation_traffic" json:"max_navigation_traffic"`
+	MaxNavigationTraffic int64     `db:"max_navigation_traffic" json:"max_navigation_traffic"`
 	MaxNavigationTime    int       `db:"max_navigation_time" json:"max_navigation_time"`
 	AutoLogin            bool      `db:"auto_login" json:"auto_login"`
 	ValidFrom            time.Time `db:"valid_from" json:"valid_from"`
